Refactor MonthlyReport and Report components to streamline date selection and loading logic. Removed unused state and improved error handling. Added month and year selection for monthly reports with appropriate options. Updated CSS for better layout of select elements.
