+++ src/mainAppPages/Report.js
@@ -1,8 +1,20 @@
-import { useEffect, useState } from "react";
+import { useEffect, useMemo, useState } from "react";
+import { Select } from "antd";
 import { useSelector } from "react-redux";
 import SendReportPanel from "../components/SendReportPanel";
 import MonthlyReport from "../components/MonthlyReport ";
 import ReportIframePreview from "../components/ReportIframePreview";
+
+const { Option } = Select;
+
+const getPreviousMonth = () => {
+  const date = new Date();
+  date.setMonth(date.getMonth() - 1);
+  return {
+    month: date.getMonth() + 1,
+    year: date.getFullYear(),
+  };
+};
 
 const Report = () => {
   const sideBar = useSelector((state) => state.sideBar);
@@ -13,11 +25,52 @@ const Report = () => {
   const [date, setDate] = useState("");
   const [startDate, setStartDate] = useState("");
   const [endDate, setEndDate] = useState("");
+  const [selectedDate, setSelectedDate] = useState(getPreviousMonth);
+
+  const currentDate = new Date();
+  const currentMonth = currentDate.getMonth() + 1;
+  const currentYear = currentDate.getFullYear();
+
+  const monthOptions = useMemo(
+    () =>
+      [
+        { value: 1, label: "January" },
+        { value: 2, label: "February" },
+        { value: 3, label: "March" },
+        { value: 4, label: "April" },
+        { value: 5, label: "May" },
+        { value: 6, label: "June" },
+        { value: 7, label: "July" },
+        { value: 8, label: "August" },
+        { value: 9, label: "September" },
+        { value: 10, label: "October" },
+        { value: 11, label: "November" },
+        { value: 12, label: "December" },
+      ].map((option) => ({
+        ...option,
+        disabled:
+          selectedDate.year === currentYear && option.value > currentMonth,
+      })),
+    [selectedDate.year, currentYear, currentMonth]
+  );
+
+  const yearOptions = useMemo(
+    () =>
+      Array.from({ length: 8 }, (_, i) => {
+        const year = currentYear - i;
+        return {
+          value: year,
+          label: year.toString(),
+          disabled: year > currentYear,
+        };
+      }),
+    [currentYear]
+  );
 
   // ✅ SINGLE SOURCE OF CONTEXT
   const [reportContext, setReportContext] = useState(null);
 
-  // ✅ DAILY + PERIODIC AUTO CONTEXT
+  // ✅ DAILY + PERIODIC + MONTHLY AUTO CONTEXT
   useEffect(() => {
     if (!branchId) {
       setReportContext(null);
@@ -40,7 +93,16 @@ const Report = () => {
         end_date: endDate,
       });
     }
-  }, [reportType, branchId, date, startDate, endDate]);
+
+    if (reportType === "monthly" && selectedDate.month && selectedDate.year) {
+      setReportContext({
+        report_type: "monthly",
+        branch_id: branchId,
+        month: selectedDate.month,
+        year: selectedDate.year,
+      });
+    }
+  }, [reportType, branchId, date, startDate, endDate, selectedDate]);
 
   return (
     <div className="report-page">
@@ -109,6 +171,59 @@ const Report = () => {
             </div>
           </div>
         )}
+
+        {reportType === "monthly" && (
+          <div className="report-page__date-range">
+            <div className="report-page__date-group">
+              <label htmlFor="report-monthly-year" className="report-page__label">
+                Year
+              </label>
+              <Select
+                id="report-monthly-year"
+                className="report-page__select"
+                value={selectedDate.year}
+                onChange={(value) =>
+                  setSelectedDate((prev) => ({ ...prev, year: value }))
+                }
+                placeholder="Select Year"
+              >
+                {yearOptions.map((option) => (
+                  <Option
+                    key={option.value}
+                    value={option.value}
+                    disabled={option.disabled}
+                  >
+                    {option.label}
+                  </Option>
+                ))}
+              </Select>
+            </div>
+            <div className="report-page__date-group">
+              <label htmlFor="report-monthly-month" className="report-page__label">
+                Month
+              </label>
+              <Select
+                id="report-monthly-month"
+                className="report-page__select"
+                value={selectedDate.month}
+                onChange={(value) =>
+                  setSelectedDate((prev) => ({ ...prev, month: value }))
+                }
+                placeholder="Select Month"
+              >
+                {monthOptions.map((option) => (
+                  <Option
+                    key={option.value}
+                    value={option.value}
+                    disabled={option.disabled}
+                  >
+                    {option.label}
+                  </Option>
+                ))}
+              </Select>
+            </div>
+          </div>
+        )}
       </section>
 
       {reportType !== "monthly" && reportContext && (
@@ -120,8 +235,8 @@ const Report = () => {
       {reportType === "monthly" && (
         <section className="report-page__monthly" aria-label="Monthly report">
           <MonthlyReport
-            branchId={branchId}
-            setReportContext={setReportContext}
+            month={selectedDate.month}
+            year={selectedDate.year}
           />
         </section>
       )}
